requirements.txt
opencv-python < 4
opencv-contrib-python < 4
numpy
pillow
